refactor & jsdoc type
